Rename WHY2025 to Tanmatsu
