+++ targets/tanmatsu/badge_bsp_device.c
@@ -1,4 +1,4 @@
-// Board support package API: WHY2025 implementation
+// Board support package API: Tanmatsu implementation
 // SPDX-FileCopyrightText: 2024 Nicolai Electronics
 // SPDX-FileCopyrightText: 2024 Orange-Murker
 // SPDX-License-Identifier: MIT
@@ -7,12 +7,12 @@
 #include "bsp/display.h"
 #include "bsp/i2c.h"
 #include "bsp/input.h"
-#include "bsp/why2025.h"
+#include "bsp/tanmatsu.h"
 #include "esp_check.h"
 #include "esp_err.h"
 #include "esp_log.h"
 #include "tanmatsu_coprocessor.h"
-#include "why2025_hardware.h"
+#include "tanmatsu_hardware.h"
 
 #include <stdbool.h>
 #include <stdint.h>
@@ -26,10 +26,10 @@ static SemaphoreHandle_t             i2c_concurrency_semaphore       = NULL;
 static tanmatsu_coprocessor_handle_t coprocessor_handle              = NULL;
 static bool                          initialized_without_coprocessor = false;
 
-static char const device_name[]         = "WHY2025 badge";
-static char const device_manufacturer[] = "Badge.Team";
+static char const device_name[]         = "Tanmatsu";
+static char const device_manufacturer[] = "Nicolai Electronics";
 
-esp_err_t bsp_why2025_coprocessor_get_handle(tanmatsu_coprocessor_handle_t *handle) {
+esp_err_t bsp_tanmatsu_coprocessor_get_handle(tanmatsu_coprocessor_handle_t *handle) {
     if (coprocessor_handle == NULL) {
         return ESP_FAIL;
     }
