+++ targets/tanmatsu/badge_bsp_power.c
@@ -1,10 +1,10 @@
-// Board support package API: WHY2025 implementation
+// Board support package API: Tanmatsu implementation
 // SPDX-FileCopyrightText: 2024 Orange-Murker
 // SPDX-FileCopyrightText: 2024 Nicolai Electronics
 // SPDX-License-Identifier: MIT
 
 #include "bsp/power.h"
-#include "bsp/why2025.h"
+#include "bsp/tanmatsu.h"
 #include "esp_check.h"
 #include "esp_err.h"
 #include "tanmatsu_coprocessor.h"
@@ -24,7 +24,7 @@ esp_err_t bsp_power_get_battery_information(bsp_power_battery_information_t *out
     ESP_RETURN_ON_FALSE(out_information, ESP_ERR_INVALID_ARG, TAG, "Information output argument is NULL");
 
     tanmatsu_coprocessor_handle_t handle = NULL;
-    ESP_RETURN_ON_ERROR(bsp_why2025_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
+    ESP_RETURN_ON_ERROR(bsp_tanmatsu_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
 
     bool    battery_attached;
     bool    charging_disabled;
@@ -74,7 +74,7 @@ esp_err_t bsp_power_get_battery_information(bsp_power_battery_information_t *out
 esp_err_t bsp_power_get_battery_voltage(uint16_t *out_millivolt) {
     ESP_RETURN_ON_FALSE(out_millivolt, ESP_ERR_INVALID_ARG, TAG, "Millivolt output argument is NULL");
     tanmatsu_coprocessor_handle_t handle = NULL;
-    ESP_RETURN_ON_ERROR(bsp_why2025_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
+    ESP_RETURN_ON_ERROR(bsp_tanmatsu_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
     ESP_RETURN_ON_ERROR(tanmatsu_coprocessor_get_pmic_vbat(handle, out_millivolt), TAG, "Failed to read battery voltage");
     return ESP_OK;
 }
@@ -82,7 +82,7 @@ esp_err_t bsp_power_get_battery_voltage(uint16_t *out_millivolt) {
 esp_err_t bsp_power_get_system_voltage(uint16_t *out_millivolt) {
     ESP_RETURN_ON_FALSE(out_millivolt, ESP_ERR_INVALID_ARG, TAG, "Millivolt output argument is NULL");
     tanmatsu_coprocessor_handle_t handle = NULL;
-    ESP_RETURN_ON_ERROR(bsp_why2025_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
+    ESP_RETURN_ON_ERROR(bsp_tanmatsu_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
     ESP_RETURN_ON_ERROR(tanmatsu_coprocessor_get_pmic_vsys(handle, out_millivolt), TAG, "Failed to read system voltage");
     return ESP_OK;
 }
@@ -90,7 +90,7 @@ esp_err_t bsp_power_get_system_voltage(uint16_t *out_millivolt) {
 esp_err_t bsp_power_get_input_voltage(uint16_t *out_millivolt) {
     ESP_RETURN_ON_FALSE(out_millivolt, ESP_ERR_INVALID_ARG, TAG, "Millivolt output argument is NULL");
     tanmatsu_coprocessor_handle_t handle = NULL;
-    ESP_RETURN_ON_ERROR(bsp_why2025_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
+    ESP_RETURN_ON_ERROR(bsp_tanmatsu_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
     ESP_RETURN_ON_ERROR(tanmatsu_coprocessor_get_pmic_vbus(handle, out_millivolt), TAG, "Failed to read input voltage");
     return ESP_OK;
 }
@@ -99,7 +99,7 @@ esp_err_t bsp_power_get_charging_configuration(bool *out_disabled, uint16_t *out
     bool                          disabled;
     uint8_t                       chrg_speed;
     tanmatsu_coprocessor_handle_t handle = NULL;
-    ESP_RETURN_ON_ERROR(bsp_why2025_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
+    ESP_RETURN_ON_ERROR(bsp_tanmatsu_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
     ESP_RETURN_ON_ERROR(tanmatsu_coprocessor_get_pmic_charging_control(handle, &disabled, &chrg_speed), TAG, "Failed to get charging configuration");
     if (out_disabled) {
         *out_disabled = disabled;
@@ -128,7 +128,7 @@ esp_err_t bsp_power_configure_charging(bool disable, uint16_t current) {
         chrg_speed = 3;
     }
     tanmatsu_coprocessor_handle_t handle = NULL;
-    ESP_RETURN_ON_ERROR(bsp_why2025_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
+    ESP_RETURN_ON_ERROR(bsp_tanmatsu_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
     ESP_RETURN_ON_ERROR(tanmatsu_coprocessor_set_pmic_charging_control(handle, disable, chrg_speed), TAG, "Failed to configure charging");
     return ESP_OK;
 }
@@ -136,14 +136,14 @@ esp_err_t bsp_power_configure_charging(bool disable, uint16_t current) {
 /*esp_err_t bsp_power_get_usb_host_boost_enabled(bool *out_enabled) {
     ESP_RETURN_ON_FALSE(out_enabled, ESP_ERR_INVALID_ARG, TAG, "Enabled output argument is NULL");
     tanmatsu_coprocessor_handle_t handle = NULL;
-    ESP_RETURN_ON_ERROR(bsp_why2025_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
+    ESP_RETURN_ON_ERROR(bsp_tanmatsu_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
     ESP_RETURN_ON_ERROR(tanmatsu_coprocessor_get_pmic_otg_control(handle, out_enabled), TAG, "Failed to get USB host boot status");
     return ESP_OK;
 }*/
 
 esp_err_t bsp_power_set_usb_host_boost_enabled(bool enable) {
     tanmatsu_coprocessor_handle_t handle = NULL;
-    ESP_RETURN_ON_ERROR(bsp_why2025_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
+    ESP_RETURN_ON_ERROR(bsp_tanmatsu_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
     ESP_RETURN_ON_ERROR(tanmatsu_coprocessor_set_pmic_otg_control(handle, enable), TAG, "Failed to set USB host boot configuration");
     return ESP_OK;
 }
@@ -151,7 +151,7 @@ esp_err_t bsp_power_set_usb_host_boost_enabled(bool enable) {
 esp_err_t bsp_power_get_radio_enabled(bool *out_enabled) {
     ESP_RETURN_ON_FALSE(out_enabled, ESP_ERR_INVALID_ARG, TAG, "Enabled output argument is NULL");
     tanmatsu_coprocessor_handle_t handle = NULL;
-    ESP_RETURN_ON_ERROR(bsp_why2025_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
+    ESP_RETURN_ON_ERROR(bsp_tanmatsu_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
     tanmatsu_coprocessor_radio_state_t status;
     ESP_RETURN_ON_ERROR(tanmatsu_coprocessor_get_radio_state(handle, &status), TAG, "Failed to get radio status");
     *out_enabled = (status == tanmatsu_coprocessor_radio_state_enabled_application);
@@ -160,7 +160,7 @@ esp_err_t bsp_power_get_radio_enabled(bool *out_enabled) {
 
 esp_err_t bsp_power_set_radio_enabled(bool enable) {
     tanmatsu_coprocessor_handle_t handle = NULL;
-    ESP_RETURN_ON_ERROR(bsp_why2025_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
+    ESP_RETURN_ON_ERROR(bsp_tanmatsu_coprocessor_get_handle(&handle), TAG, "Failed to get coprocessor handle");
     ESP_RETURN_ON_ERROR(
         tanmatsu_coprocessor_set_radio_state(handle, enable ? tanmatsu_coprocessor_radio_state_enabled_application : tanmatsu_coprocessor_radio_state_disabled),
         TAG,
